requirements.txt
accelerate==0.28.0
diffusers==0.27.2
gradio==4.26.0
invisible-watermark==0.2.0
Pillow==10.3.0
spaces==0.26.0
torch==2.0.0
transformers==4.39.3
flask
flask-cord
asgiref
